change GTP context next idle/busy state logic
A GTP context is busy when is waits for some thing to occure between
handle_event call, e.g. for the reply to a reuest is has sent.
Unconditionally swithcing to idle is wrong is most cases, so change
that.

Note: we might want to change the simple init/idle/busy logic with
a list (or queue) of pending/outstanding actions/replies in the future.